Add test for all posts
