+++ spec/views/post/show.html.erb_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+RSpec.describe 'user/show.html.erb', type: :view do
+  pending "add some examples to (or delete) #{__FILE__}"
+end
